+++ otelconf/v0.3.0/trace_test.go
@@ -4,11 +4,14 @@
 package otelconf
 
 import (
+	"bytes"
 	"context"
 	"crypto/tls"
 	"crypto/x509"
 	"errors"
 	"net"
+	"net/http"
+	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -22,6 +25,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 
+	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
 	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace"
@@ -33,7 +37,7 @@ import (
 	v1 "go.opentelemetry.io/proto/otlp/collector/trace/v1"
 )
 
-func TestTracerPovider(t *testing.T) {
+func TestTracerProvider(t *testing.T) {
 	tests := []struct {
 		name         string
 		cfg          configOptions
@@ -114,6 +118,58 @@ func TestTracerPovider(t *testing.T) {
 		assert.Equal(t, tt.wantErr, err)
 		require.NoError(t, shutdown(context.Background()))
 	}
+}
+
+func TestTracerProviderOptions(t *testing.T) {
+	var calls int
+	srv := httptest.NewServer(http.HandlerFunc(func(http.ResponseWriter, *http.Request) {
+		calls++
+	}))
+	defer srv.Close()
+
+	cfg := OpenTelemetryConfiguration{
+		TracerProvider: &TracerProvider{
+			Processors: []SpanProcessor{{
+				Simple: &SimpleSpanProcessor{
+					Exporter: SpanExporter{
+						OTLP: &OTLP{
+							Protocol: ptr("http/protobuf"),
+							Endpoint: ptr(srv.URL),
+							Insecure: ptr(true),
+						},
+					},
+				},
+			}},
+		},
+	}
+
+	var buf bytes.Buffer
+	stdouttraceExporter, err := stdouttrace.New(stdouttrace.WithWriter(&buf))
+	require.NoError(t, err)
+
+	res := resource.NewSchemaless(attribute.String("foo", "bar"))
+	sdk, err := NewSDK(
+		WithOpenTelemetryConfiguration(cfg),
+		WithTracerProviderOptions(sdktrace.WithSyncer(stdouttraceExporter)),
+		WithTracerProviderOptions(sdktrace.WithResource(res)),
+	)
+	require.NoError(t, err)
+	defer func() {
+		assert.NoError(t, sdk.Shutdown(context.Background()))
+	}()
+
+	// The exporter, which we passed in as an extra option to NewSDK,
+	// should be wired up to the provider in addition to the
+	// configuration-based OTLP exporter.
+	tracer := sdk.TracerProvider().Tracer("test")
+	_, span := tracer.Start(context.Background(), "span")
+	span.End()
+	assert.NotZero(t, buf)
+	assert.Equal(t, 1, calls)
+	// Options provided by WithMeterProviderOptions may be overridden
+	// by configuration, e.g. the resource is always defined via
+	// configuration.
+	assert.NotContains(t, buf.String(), "foo")
 }
 
 func TestSpanProcessor(t *testing.T) {